adjusted login and signup pages
